+++ FinanceManager.Tests/Reports/ReportAggregationServiceAdditionalTests.cs
@@ -53,7 +53,7 @@ public sealed class ReportAggregationServiceAdditionalTests
         await db.SaveChangesAsync();
 
         var sut = new ReportAggregationService(db);
-        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, (int)PostingKind.Contact, ReportInterval.Month, 12, IncludeCategory: false, ComparePrevious: true, CompareYear: false, AnalysisDate: new DateTime(2025,2,1)), CancellationToken.None);
+        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, PostingKind.Contact, ReportInterval.Month, 12, IncludeCategory: false, ComparePrevious: true, CompareYear: false, AnalysisDate: new DateTime(2025,2,1)), CancellationToken.None);
 
         Assert.All(result.Points, p => Assert.StartsWith("Contact:", p.GroupKey));
         Assert.DoesNotContain(result.Points, p => p.GroupKey.StartsWith("Category:"));
@@ -81,7 +81,7 @@ public sealed class ReportAggregationServiceAdditionalTests
 
         var sut = new ReportAggregationService(db);
         var latest = new DateTime(2025,2,1);
-        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, (int)PostingKind.Contact, ReportInterval.Month, 12, IncludeCategory: true, ComparePrevious: true, CompareYear: false, AnalysisDate: latest), CancellationToken.None);
+        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, PostingKind.Contact, ReportInterval.Month, 12, IncludeCategory: true, ComparePrevious: true, CompareYear: false, AnalysisDate: latest), CancellationToken.None);
 
         // c2 should have auto-added zero row in Feb with previous = Jan amount
         var c2Feb = result.Points.SingleOrDefault(p => p.GroupKey == $"Contact:{c2.Id}" && p.PeriodStart == latest);
@@ -106,7 +106,7 @@ public sealed class ReportAggregationServiceAdditionalTests
         await db.SaveChangesAsync();
 
         var sut = new ReportAggregationService(db);
-        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, (int)PostingKind.Contact, ReportInterval.Month, 12, IncludeCategory: false, ComparePrevious: true, CompareYear: true, AnalysisDate: new DateTime(2025,2,1)), CancellationToken.None);
+        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, PostingKind.Contact, ReportInterval.Month, 12, IncludeCategory: false, ComparePrevious: true, CompareYear: true, AnalysisDate: new DateTime(2025,2,1)), CancellationToken.None);
 
         // Group for c1 should be removed (only zero data + zero row + no previous/year non-zero)
         Assert.DoesNotContain(result.Points, p => p.GroupKey == $"Contact:{c1.Id}");
@@ -134,7 +134,7 @@ public sealed class ReportAggregationServiceAdditionalTests
 
         var sut = new ReportAggregationService(db);
         var take = 5;
-        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, (int)PostingKind.Contact, ReportInterval.Month, take, IncludeCategory: false, ComparePrevious: false, CompareYear: false, AnalysisDate: new DateTime(2025,3,1)), CancellationToken.None);
+        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, PostingKind.Contact, ReportInterval.Month, take, IncludeCategory: false, ComparePrevious: false, CompareYear: false, AnalysisDate: new DateTime(2025,3,1)), CancellationToken.None);
 
         var periods = result.Points.Select(p => p.PeriodStart).Distinct().OrderBy(d=>d).ToList();
         Assert.Equal(take, periods.Count);
@@ -162,17 +162,17 @@ public sealed class ReportAggregationServiceAdditionalTests
 
         var sut = new ReportAggregationService(db);
 
-        var quarters = await sut.QueryAsync(new ReportAggregationQuery(user.Id, (int)PostingKind.Contact, ReportInterval.Quarter, 10, IncludeCategory: false, ComparePrevious: true, CompareYear: true, AnalysisDate: new DateTime(2024,4,1)), CancellationToken.None);
+        var quarters = await sut.QueryAsync(new ReportAggregationQuery(user.Id, PostingKind.Contact, ReportInterval.Quarter, 10, IncludeCategory: false, ComparePrevious: true, CompareYear: true, AnalysisDate: new DateTime(2024,4,1)), CancellationToken.None);
         Assert.Equal(2, quarters.Points.Count(p=>p.GroupKey.StartsWith("Contact:")));
         var q2Point = quarters.Points.Single(p=>p.PeriodStart == new DateTime(2024,4,1) && p.GroupKey.StartsWith("Contact:"));
         Assert.Equal(100m, q2Point.PreviousAmount);
 
-        var halfYears = await sut.QueryAsync(new ReportAggregationQuery(user.Id, (int)PostingKind.Contact, ReportInterval.HalfYear, 10, IncludeCategory: false, ComparePrevious: true, CompareYear: false, AnalysisDate: new DateTime(2024,7,1)), CancellationToken.None);
+        var halfYears = await sut.QueryAsync(new ReportAggregationQuery(user.Id, PostingKind.Contact, ReportInterval.HalfYear, 10, IncludeCategory: false, ComparePrevious: true, CompareYear: false, AnalysisDate: new DateTime(2024,7,1)), CancellationToken.None);
         Assert.Equal(2, halfYears.Points.Count);
         var h2Point = halfYears.Points.Single(p=>p.PeriodStart == new DateTime(2024,7,1));
         Assert.Equal(250m, h2Point.PreviousAmount);
 
-        var years = await sut.QueryAsync(new ReportAggregationQuery(user.Id, (int)PostingKind.Contact, ReportInterval.Year, 10, IncludeCategory: false, ComparePrevious: true, CompareYear: true, AnalysisDate: new DateTime(2025,1,1)), CancellationToken.None);
+        var years = await sut.QueryAsync(new ReportAggregationQuery(user.Id, PostingKind.Contact, ReportInterval.Year, 10, IncludeCategory: false, ComparePrevious: true, CompareYear: true, AnalysisDate: new DateTime(2025,1,1)), CancellationToken.None);
         Assert.Equal(2, years.Points.Count);
         var y2025Point = years.Points.Single(p=>p.PeriodStart == new DateTime(2025,1,1));
         Assert.Equal(550m, y2025Point.PreviousAmount);
@@ -191,7 +191,7 @@ public sealed class ReportAggregationServiceAdditionalTests
         db.PostingAggregates.Add(agg);
         await db.SaveChangesAsync();
         var sut = new ReportAggregationService(db);
-        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, (int)PostingKind.Contact, ReportInterval.Month, 5, IncludeCategory: true, ComparePrevious: false, CompareYear: false, AnalysisDate: new DateTime(2025,3,1)), CancellationToken.None);
+        var result = await sut.QueryAsync(new ReportAggregationQuery(user.Id, PostingKind.Contact, ReportInterval.Month, 5, IncludeCategory: true, ComparePrevious: false, CompareYear: false, AnalysisDate: new DateTime(2025,3,1)), CancellationToken.None);
         Assert.Contains(result.Points, p => p.GroupKey == $"Category:{PostingKind.Contact}:_none" && p.Amount == 42m);
         var child = result.Points.Single(p => p.GroupKey == $"Contact:{c1.Id}");
         Assert.Equal($"Category:{PostingKind.Contact}:_none", child.ParentGroupKey);
